Add REDT for extension info
Different RISC-V platforms can support different extensions. A new ACPI table is required to communicate this information to the OS.

This table will be maintained by the RVI and a reference will be added in ACPI spec. This table will be made mandatory in the RISC-V ACPI requirements document.

Any new extension should reserve its unique ID (16 bit) if ACPI support is required for the platform implementing the extension.

Signed-off-by: Sunil V L <sunilvl@ventanamicro.com>