Add RTDT
Add RISC-V Timer Description Table (RTDT).

This table is required to communicate information regarding timer devices. Currently, only time_base_frequency is added. But in future revisions of this table, more information regarding timer device and watchdog device can be added to this table.

The plan is to maintain this table within RVI organization and provide a reference to it from the ACPI spec. A separate code-first proposal may be required to add the reference in ACPI spec.

Signed-off-by: Sunil V L <sunilvl@ventanamicro.com>