RIMT: Simplify fields referring to a number of items
I don't think having a field referring to a number of items should
include the complexity of the so called N-1 convention.

If the goal is to enforce these fields to never be 0, then let's write
it down in the specification so that different implementations can error
out if these fields hold a value of 0.

Signed-off-by: Sebastien Boeuf <seb@rivosinc.com>